move  component/constants in to core/constants
